requirements.txt
openai>=1.40.0
google-search-results>=2.4.2
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
